Added Date Conditional and closing </td>
I added a conditional to show the correct dates in the classified ads panel. Previously, if a post had not been confirmed/published it was given a year of 1969. This will check if the post is published, if so it renders the published date, otherwise, it renders the date it was created.

Also added a closing </td>